refactor: Move Panels to layout folder
