+++ plugins/ui/src/js/src/layout/ReactPanelManager.ts
