+++ plugins/ui/src/js/src/layout/Row.tsx
@@ -3,7 +3,7 @@ import { useLayoutManager } from '@deephaven/dashboard';
 import type { RowOrColumn } from '@deephaven/golden-layout';
 import { normalizeRowChildren, type RowElementProps } from './LayoutUtils';
 import { ParentItemContext, useParentItem } from './ParentItemContext';
-import { usePanelId } from '../ReactPanelContext';
+import { usePanelId } from './ReactPanelContext';
 import { Flex } from '../elements/spectrum';
 
 function LayoutRow({ children, height }: RowElementProps): JSX.Element | null {
